requirements.txt
crewai
crewai-tools
litellm
streamlit
requests
python-dotenv
pydantic
tenacity
PyYAML
